fix(posts-cr): apply Copilot CR feedback on security and code quality
- PostContent: narrow rehype-sanitize style attribute from * to span/svg only
  (KaTeX only needs style on these two elements; global style is an XSS vector)
- Extract buildFrontmatter to lib/frontmatter.ts to avoid pulling editor
  bundle into detail page / card bundles
- PostDetailOwnerActions: add .catch(()=>({})) on DELETE res.json() for
  resilient error body parsing
- EditorPageClient: align titleToSlug comment with actual Unicode behavior,
  add tags trim+filter before POST, guard satoken header to avoid empty token
- PromoteToDocsButton: update import path, guard satoken header
